requirements.txt
absl-py
asttokens
astunparse
beautifulsoup4
cachetools
certifi
charset-normalizer
comm
contourpy
cycler
db-dtypes
debugpy
decorator
executing
flatbuffers
fonttools
gast
google
google-api-core
google-auth
google-cloud-bigquery
google-cloud-core
google-cloud-storage
google-crc32c
google-pasta
google-resumable-media
googleapis-common-protos
grpcio
grpcio-status
h5py
idna
ipykernel
ipython
jedi
joblib
jupyter_client
jupyter_core
keras
kiwisolver
libclang
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
ml-dtypes
namex
nest-asyncio
numpy
opt-einsum
optree
packaging
pandas
parso
pexpect
pillow
platformdirs
prompt_toolkit
proto-plus
protobuf
psutil
ptyprocess
pure_eval
pyarrow
pyasn1
pyasn1_modules
pygeohash
Pygments
pyparsing
python-dateutil
python-dotenv
pytz
pyzmq
requests
rich
rsa
scikit-learn
scipy
seaborn
six
soupsieve
stack-data
termcolor
threadpoolctl
tornado
traitlets
typing_extensions
tzdata
urllib3
wcwidth
Werkzeug
wrapt
